Create P3Rev2merged.c
I merged your file with mine so we can work on same file. I didn't remove anything but added from my code. This will help see whatchanged in git, I hope. If you run it, you can see that it skips instruction 17 in ID and IF stages but not in EX.